fix: rename the package identifier to solve eas build error
