+++ android/src/main/AndroidManifest.xml
@@ -1,3 +1,3 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-          package="so.onekey.app.wallet">
+          package="so.onekey.app.wallet.lite">
 </manifest>
