+++ android/src/main/java/so/onekey/app/wallet/lite/OKLitePackage.kt
@@ -1,0 +1,16 @@
+package so.onekey.app.wallet.lite
+
+import com.facebook.react.ReactPackage
+import com.facebook.react.bridge.NativeModule
+import com.facebook.react.bridge.ReactApplicationContext
+import com.facebook.react.uimanager.ViewManager
+
+class OKLitePackage : ReactPackage {
+    override fun createNativeModules(reactContext: ReactApplicationContext): List<NativeModule> =
+        listOf<NativeModule>(
+            OKLiteManager(reactContext),
+        )
+
+    override fun createViewManagers(reactContext: ReactApplicationContext) =
+        emptyList<ViewManager<*, *>>()
+}
