+++ android/src/main/java/so/onekey/app/wallet/lite/keys/KeysNativeProvider.kt
@@ -1,0 +1,13 @@
+package so.onekey.app.wallet.lite.keys
+
+import android.content.Context
+
+class KeysNativeProvider {
+    companion object {
+        init {
+            System.loadLibrary("keys")
+        }
+    }
+
+    external fun getLiteSecureChannelInitParams(context: Context): String
+}
