+++ android/src/main/java/so/onekey/app/wallet/lite/utils/Utils.java
@@ -1,0 +1,433 @@
+package so.onekey.app.wallet.lite.utils;
+
+import android.annotation.SuppressLint;
+import android.app.Activity;
+import android.app.ActivityManager;
+import android.app.Application;
+import android.app.Application.ActivityLifecycleCallbacks;
+import android.content.Context;
+import android.os.Bundle;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+
+import com.facebook.react.bridge.ReactApplicationContext;
+
+import org.jetbrains.annotations.NotNull;
+
+import java.lang.reflect.Field;
+import java.lang.reflect.InvocationTargetException;
+import java.util.HashMap;
+import java.util.LinkedList;
+import java.util.List;
+import java.util.Map;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
+
+/**
+ * <pre>
+ *     author:
+ *                                      ___           ___           ___         ___
+ *         _____                       /  /\         /__/\         /__/|       /  /\
+ *        /  /::\                     /  /::\        \  \:\       |  |:|      /  /:/
+ *       /  /:/\:\    ___     ___    /  /:/\:\        \  \:\      |  |:|     /__/::\
+ *      /  /:/~/::\  /__/\   /  /\  /  /:/~/::\   _____\__\:\   __|  |:|     \__\/\:\
+ *     /__/:/ /:/\:| \  \:\ /  /:/ /__/:/ /:/\:\ /__/::::::::\ /__/\_|:|____    \  \:\
+ *     \  \:\/:/~/:/  \  \:\  /:/  \  \:\/:/__\/ \  \:\~~\~~\/ \  \:\/:::::/     \__\:\
+ *      \  \::/ /:/    \  \:\/:/    \  \::/       \  \:\  ~~~   \  \::/~~~~      /  /:/
+ *       \  \:\/:/      \  \::/      \  \:\        \  \:\        \  \:\         /__/:/
+ *        \  \::/        \__\/        \  \:\        \  \:\        \  \:\        \__\/
+ *         \__\/                       \__\/         \__\/         \__\/
+ *     blog  : http://blankj.com
+ *     time  : 16/12/08
+ *     desc  : utils about initialization
+ * </pre>
+ */
+public final class Utils {
+
+    @SuppressLint("StaticFieldLeak")
+    private static Application sApplication;
+
+    static final ActivityLifecycleImpl ACTIVITY_LIFECYCLE = new ActivityLifecycleImpl();
+
+    private Utils() {
+        throw new UnsupportedOperationException("u can't instantiate me...");
+    }
+
+    /**
+     * Init utils.
+     *
+     * <p>Init it in the class of Application.
+     *
+     * @param context context
+     */
+    public static void init(@NonNull final ReactApplicationContext context) {
+        Utils.getActivityLifecycle().setTopActivity(context.getCurrentActivity());
+        init((Application) context.getApplicationContext());
+    }
+
+    /**
+     * Init utils.
+     *
+     * <p>Init it in the class of Application.
+     *
+     * @param app application
+     */
+    public static void init(@NonNull final Application app) {
+        if (sApplication == null) {
+            Utils.sApplication = app;
+            Utils.sApplication.registerActivityLifecycleCallbacks(ACTIVITY_LIFECYCLE);
+        }
+    }
+
+    /**
+     * Return the context of Application object.
+     *
+     * @return the context of Application object
+     */
+    public static Application getApp() {
+        if (sApplication != null) {
+            return sApplication;
+        }
+        try {
+            @SuppressLint("PrivateApi")
+            Class<?> activityThread = Class.forName("android.app.ActivityThread");
+            Object at = activityThread.getMethod("currentActivityThread").invoke(null);
+            Object app = activityThread.getMethod("getApplication").invoke(at);
+            if (app == null) {
+                throw new NullPointerException("u should init first");
+            }
+            init((Application) app);
+            return sApplication;
+        } catch (NoSuchMethodException e) {
+            e.printStackTrace();
+        } catch (IllegalAccessException e) {
+            e.printStackTrace();
+        } catch (InvocationTargetException e) {
+            e.printStackTrace();
+        } catch (ClassNotFoundException e) {
+            e.printStackTrace();
+        }
+        throw new NullPointerException("u should init first");
+    }
+
+    public static ActivityLifecycleImpl getActivityLifecycle() {
+        return ACTIVITY_LIFECYCLE;
+    }
+
+    public static LinkedList<Activity> getActivityList() {
+        return ACTIVITY_LIFECYCLE.mActivityList;
+    }
+
+    @NotNull
+    public static Context getTopActivityOrApp() {
+        if (isAppForeground()) {
+            Activity topActivity = ACTIVITY_LIFECYCLE.getTopActivity();
+            return topActivity == null ? Utils.getApp() : topActivity;
+        } else {
+            return Utils.getApp();
+        }
+    }
+
+    @Nullable
+    public static Activity getTopActivity() {
+        if (!isAppForeground()) {
+            return null;
+        }
+        return ACTIVITY_LIFECYCLE.getTopActivity();
+    }
+
+    /**
+     * Finish all of activities.
+     */
+    public static void finishAllActivities() {
+        finishAllActivities(false);
+    }
+
+    /**
+     * Finish all of activities.
+     *
+     * @param isLoadAnim True to use animation for the outgoing activity, false otherwise.
+     */
+    public static void finishAllActivities(final boolean isLoadAnim) {
+        List<Activity> activityList = getActivityList();
+        for (Activity act : activityList) {
+            // sActivityList remove the index activity at onActivityDestroyed
+            act.finish();
+            if (!isLoadAnim) {
+                act.overridePendingTransition(0, 0);
+            }
+        }
+    }
+
+    /**
+     * Finish all of activities except the newest activity.
+     */
+    public static void finishAllActivitiesExceptNewest() {
+        finishAllActivitiesExceptNewest(false);
+    }
+
+    /**
+     * Finish all of activities except the newest activity.
+     *
+     * @param isLoadAnim True to use animation for the outgoing activity, false otherwise.
+     */
+    public static void finishAllActivitiesExceptNewest(final boolean isLoadAnim) {
+        List<Activity> activities = getActivityList();
+        for (int i = 1; i < activities.size(); i++) {
+            finishActivity(activities.get(i), isLoadAnim);
+        }
+    }
+
+    public static void finishActivity(@NonNull final Activity activity) {
+        finishActivity(activity, false);
+    }
+
+    public static void finishActivity(@NonNull final Activity activity, final boolean isLoadAnim) {
+        activity.finish();
+        if (!isLoadAnim) {
+            activity.overridePendingTransition(0, 0);
+        }
+    }
+
+    /**
+     * Finish the activity.
+     *
+     * @param clz The activity class.
+     */
+    public static void finishActivity(@NonNull final Class<? extends Activity> clz) {
+        finishActivity(clz, false);
+    }
+
+    /**
+     * Finish the activity.
+     *
+     * @param clz        The activity class.
+     * @param isLoadAnim True to use animation for the outgoing activity, false otherwise.
+     */
+    public static void finishActivity(
+            @NonNull final Class<? extends Activity> clz, final boolean isLoadAnim) {
+        List<Activity> activities = getActivityList();
+        for (Activity activity : activities) {
+            if (activity.getClass().equals(clz)) {
+                activity.finish();
+                if (!isLoadAnim) {
+                    activity.overridePendingTransition(0, 0);
+                }
+            }
+        }
+    }
+
+    /**
+     * Finish the activities whose type not equals the activity class.
+     *
+     * @param clz The activity class.
+     */
+    public static void finishOtherActivities(@NonNull final Class<? extends Activity> clz) {
+        finishOtherActivities(clz, false);
+    }
+
+    /**
+     * Finish the activities whose type not equals the activity class.
+     *
+     * @param clz        The activity class.
+     * @param isLoadAnim True to use animation for the outgoing activity, false otherwise.
+     */
+    public static void finishOtherActivities(
+            @NonNull final Class<? extends Activity> clz, final boolean isLoadAnim) {
+        List<Activity> activities = getActivityList();
+        for (Activity act : activities) {
+            if (!act.getClass().equals(clz)) {
+                finishActivity(act, isLoadAnim);
+            }
+        }
+    }
+
+    public static boolean isAppForeground() {
+        ActivityManager am =
+                (ActivityManager) Utils.getApp().getSystemService(Context.ACTIVITY_SERVICE);
+        if (am == null) {
+            return false;
+        }
+        List<ActivityManager.RunningAppProcessInfo> info = am.getRunningAppProcesses();
+        if (info == null || info.size() == 0) {
+            return false;
+        }
+        for (ActivityManager.RunningAppProcessInfo aInfo : info) {
+            if (aInfo.importance == ActivityManager.RunningAppProcessInfo.IMPORTANCE_FOREGROUND) {
+                return aInfo.processName.equals(Utils.getApp().getPackageName());
+            }
+        }
+        return false;
+    }
+
+    static class ActivityLifecycleImpl implements ActivityLifecycleCallbacks {
+
+        final LinkedList<Activity> mActivityList = new LinkedList<>();
+        final HashMap<Object, OnAppStatusChangedListener> mStatusListenerMap = new HashMap<>();
+
+        private int mForegroundCount = 0;
+        private int mConfigCount = 0;
+
+        void addListener(final Object object, final OnAppStatusChangedListener listener) {
+            mStatusListenerMap.put(object, listener);
+        }
+
+        void removeListener(final Object object) {
+            mStatusListenerMap.remove(object);
+        }
+
+        @Override
+        public void onActivityCreated(Activity activity, Bundle savedInstanceState) {
+            setTopActivity(activity);
+        }
+
+        @Override
+        public void onActivityStarted(Activity activity) {
+            setTopActivity(activity);
+            if (mForegroundCount <= 0) {
+                postStatus(true);
+            }
+            if (mConfigCount < 0) {
+                ++mConfigCount;
+            } else {
+                ++mForegroundCount;
+            }
+        }
+
+        @Override
+        public void onActivityResumed(Activity activity) {
+            setTopActivity(activity);
+        }
+
+        @Override
+        public void onActivityPaused(Activity activity) {
+            /**/
+        }
+
+        @Override
+        public void onActivityStopped(Activity activity) {
+            if (activity.isChangingConfigurations()) {
+                --mConfigCount;
+            } else {
+                --mForegroundCount;
+                if (mForegroundCount <= 0) {
+                    postStatus(false);
+                }
+            }
+        }
+
+        @Override
+        public void onActivitySaveInstanceState(Activity activity, Bundle outState) {
+            /**/
+        }
+
+        @Override
+        public void onActivityDestroyed(Activity activity) {
+            mActivityList.remove(activity);
+        }
+
+        private void postStatus(final boolean isForeground) {
+            if (mStatusListenerMap.isEmpty()) {
+                return;
+            }
+            for (OnAppStatusChangedListener onAppStatusChangedListener :
+                    mStatusListenerMap.values()) {
+                if (onAppStatusChangedListener == null) {
+                    return;
+                }
+                if (isForeground) {
+                    onAppStatusChangedListener.onForeground();
+                } else {
+                    onAppStatusChangedListener.onBackground();
+                }
+            }
+        }
+
+        private void setTopActivity(final Activity activity) {
+            if (mActivityList.contains(activity)) {
+                if (!mActivityList.getLast().equals(activity)) {
+                    mActivityList.remove(activity);
+                    mActivityList.addLast(activity);
+                }
+            } else {
+                mActivityList.addLast(activity);
+            }
+        }
+
+        @Nullable
+        Activity getTopActivity() {
+            if (!mActivityList.isEmpty()) {
+                final Activity topActivity = mActivityList.getLast();
+                if (topActivity != null) {
+                    return topActivity;
+                }
+            }
+            // using reflect to get top activity
+            try {
+                @SuppressLint("PrivateApi")
+                Class<?> activityThreadClass = Class.forName("android.app.ActivityThread");
+                Object activityThread =
+                        activityThreadClass.getMethod("currentActivityThread").invoke(null);
+                Field activitiesField = activityThreadClass.getDeclaredField("mActivityList");
+                activitiesField.setAccessible(true);
+                Map activities = (Map) activitiesField.get(activityThread);
+                if (activities == null) {
+                    return null;
+                }
+                for (Object activityRecord : activities.values()) {
+                    Class activityRecordClass = activityRecord.getClass();
+                    Field pausedField = activityRecordClass.getDeclaredField("paused");
+                    pausedField.setAccessible(true);
+                    if (!pausedField.getBoolean(activityRecord)) {
+                        Field activityField = activityRecordClass.getDeclaredField("activity");
+                        activityField.setAccessible(true);
+                        Activity activity = (Activity) activityField.get(activityRecord);
+                        setTopActivity(activity);
+                        return activity;
+                    }
+                }
+            } catch (ClassNotFoundException e) {
+                e.printStackTrace();
+            } catch (IllegalAccessException e) {
+                e.printStackTrace();
+            } catch (InvocationTargetException e) {
+                e.printStackTrace();
+            } catch (NoSuchMethodException e) {
+                e.printStackTrace();
+            } catch (NoSuchFieldException e) {
+                e.printStackTrace();
+            }
+            return null;
+        }
+    }
+
+    ///////////////////////////////////////////////////////////////////////////
+    // interface
+    ///////////////////////////////////////////////////////////////////////////
+
+    public interface OnAppStatusChangedListener {
+
+        void onForeground();
+
+        void onBackground();
+    }
+
+    public static int dp2px(final @NotNull Context context, final Float dpValue) {
+        final float scale = context.getResources().getDisplayMetrics().density;
+        return (int) (dpValue * scale + 0.5f);
+    }
+
+    public static int sp2px(final @NotNull Context context, final Float sp) {
+        final float scale = context.getResources().getDisplayMetrics().scaledDensity;
+        return (int) (sp * scale + 0.5f);
+    }
+
+    public static boolean isEmail(String email) {
+        if (null == email || "".equals(email)) return false;
+        // Pattern p = Pattern.compile("\\w+@(\\w+.)+[a-z]{2,3}"); //简单匹配
+        Pattern p = Pattern.compile("\\w+([-+.]\\w+)*@\\w+([-.]\\w+)*\\.\\w+([-.]\\w+)*"); // 复杂匹配
+        Matcher m = p.matcher(email);
+        return m.matches();
+    }
+}
